[alpha_factory] refactor orchestrator utils
